Major overhaul of website.
Extensively modifed and fixed numerous bugs and bad formatting.
Rewrote the downloads page to be more consistent and better on all devices.
Added a patchset page, and added an entry on the front page.
Wrote two blog posts, one on the release of Dapper Linux 26 and one on the patchset.
Added new screenshots for the release of Dapper Linux 26.
Lots of polish added everywhere.